backend/requirements.txt
django==5.2
channels-redis==4.2.1
daphne==4.0.0
psycopg2-binary==2.9.10
CCXT==4.4.72
